==> scripts/pipeline/09_pipeline_SST_climatologies.qmd <==
---
title: "pipeline_SST_climatologies"
format: html
editor: visual
author: "Molly Timmers"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Load Packages

```{r}

rm(list = ls())

pkgTest <- function(x)
{
  if (!require(x,character.only = TRUE))
  {
    install.packages(x,dep=TRUE,repos='http://cran.us.r-project.org')
    if(!require(x,character.only = TRUE)) stop(x, " :Package not found")
  }
}


# Create list of required packages
list.of.packages <- c("ncdf4", "ggplot2", "sf", "dplyr", "rerddap","scales","RColorBrewer", "readxl","tidyverse", "stringr","stringi")
 

# Create list of installed packages
pkges = installed.packages()[,"Package"]

# Install and load all required pkgs
for (pk in list.of.packages) {
  pkgTest(pk)
}

```

## Data Setup

```{r}

sites<-readRDS("~/Google Drive/My Drive/Pristine Seas/SCIENCE/expeditions/PNG-2024/data/primary/processed/sites.rds")

survey_df<-sites$uvs %>%
  data.frame()


# Set ERDDAP address for Coral Reef Watch monthly Sea Surface Temperature (SST) means
erddapSST_CRW <-'https://oceanwatch.pifsc.noaa.gov/erddap/griddap/CRW_sst_v3_1_monthly.csv?sea_surface_temperature'

```

## Acquiring bleaching threshold

Selecting 3 random points per subregions to extract monthly mean SST satellite data from 1985 until 2012 following the Coral Reef Watch (CRW) standard methods for the climatological range. From this you derive the maximum monthly mean temperature (MMM) which is the highest mean monthly temperature. You calculate this by taking the means for each month within regions or subregions, whatever you choose, and then selecting the month mean with the highest temperature to represent the MMM. To access the bleaching threshold in a region/subregion, there are two methods. The CRW method is to add one degree from the MMM. However, the Kleypass method based on data from the Coral Triangle, created the bleaching threshold based on 2.5 std of the mean. We calculate both from the output. However, Kleypass does not use the climatological range of 1985-2012 rather a historical temperature range from 1960-2007 using 5-km resolution Regional Ocean Model System for the Coral Triangle (CT-ROMS) to define MMM and subsequent threshold values - Kleypass et al. 2015.

With that said, we could dig deeper and do fewer time horizons on potential bleaching as a proxy for adaption (using fewer and more recent years to calculate threshold = more rapid/recent adaption) - Logan et al. 2014.

```{r}
# Data frame of random selection of representative sites at within subregions
threshold<-survey_df %>%
    group_by(subregion) %>%
  slice_sample(n = 3) %>% # can do more but only 3 sites in Tinakula.
  data.frame()

#Creating incomplete vectors to populate extracted data Start and End dates to extract data based on CRW climatological thresholds
PNG_SST_climatology = rep(NA,11)
threshold$start<-'1985-06-01' 
threshold$end<-'2012-12-01' 


# Looping through each survey data point
for (i in 1:nrow(threshold)) {

# Create erddap URL by adding lat, lon, dates of each  point 
   url = paste(erddapSST_CRW,"[(" ,threshold$start[i], "):1:(" ,threshold$end[i], ")][(" , threshold$latitude[i],  "):1:(" ,threshold$latitude[i], ")][(" ,threshold$longitude[i], "):1:(",threshold$longitude[i], ")]",sep = "")
   
   # Request and load satelite data from ERDDAP and add corresponging
   new = read.csv(url, skip=2, header = FALSE) 
   new$ps_site_id = threshold[i, 'ps_site_id']
     new$region = threshold[i, 'region']
   new$subregion = threshold[i, 'subregion']
      new$habitat = threshold[i, 'habitat']
            new$exposure = threshold[i, 'exposure']
   new$lat = threshold[i, 'latitude']
   new$lon = threshold[i, 'longitude']
   # Append the data
   PNG_SST_climatology = rbind(PNG_SST_climatology, new)   
}

# Delete the first row (default column names)
PNG_SST_climatology = PNG_SST_climatology[-1, ]

# Rename columns
names(PNG_SST_climatology) <- c('date', 'lat_data', 'lon_data', 'SST', 'ps_site_id','region','subregion',  'habitat','exposure', 'lat', 'lon')

write.csv(PNG_SST_climatology, "PNG_climatologies.csv", row.names = F)


### getting the monthly means for the broad region across this time period.
PNG_region_thresholds<-PNG_SST_climatology %>%
  mutate(date = substr(date, 0, 10)) %>% 
  mutate(month = substr(date, 0, 7)) %>% 
mutate(month = str_sub(month,start = -2)) %>%
  group_by(region, month)%>%
  dplyr::summarise(mean = mean(SST,na.rm=T), std = sd(SST,na.rm=T) ) %>% 
  mutate(stdx2 = std*2.5) %>% 
   group_by(region) %>% 
    top_n(1,mean) %>% 
  mutate(CRW_threshold = mean + 1) %>% 
  mutate(Kleypass_threshold = mean + stdx2)

# outputting this to void repeating
write.csv(PNG_region_thresholds, "PNG_region_MMM_thresholds.csv", row.names = F)

### getting the monthly means across subregions this time period.
PNG_subregion_thresholds<-PNG_SST_climatology %>%
  mutate(date = substr(date, 0, 10)) %>% 
  mutate(month = substr(date, 0, 7)) %>% 
mutate(month = str_sub(month,start = -2)) %>%
  group_by(subregion, month)%>%
  dplyr::summarise(mean = mean(SST,na.rm=T), std = sd(SST,na.rm=T) ) %>% 
  mutate(stdx2 = std*2.5) %>% 
   group_by(subregion) %>% 
    top_n(1,mean) %>% 
  mutate(CRW_threshold = mean + 1) %>% 
  mutate(Kleypass_threshold = mean + stdx2) 

# outputting this to void repeating
write.csv(PNG_subregion_thresholds, "PNG_subregion_MMM_thresholds.csv", row.names = F)

### Saving all of the output
save(PNG_SST_climatology, PNG_region_thresholds,PNG_subregion_thresholds, file = "~/Google Drive/My Drive/Pristine Seas/SCIENCE/expeditions/PNG-2024/data/primary/processed/satellite/PNG_satellite_SST.RData")




```
